@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
+    "@angular/animations": "~14.3.0",
     "@angular/cdk": "^13.3.3",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
+    "@angular/common": "~14.3.0",
+    "@angular/compiler": "~14.3.0",
+    "@angular/core": "~14.3.0",
+    "@angular/forms": "~14.3.0",
     "@angular/material": "^13.3.5",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/platform-browser": "~14.3.0",
+    "@angular/platform-browser-dynamic": "~14.3.0",
+    "@angular/router": "~14.3.0",
     "@syncfusion/ej2-angular-richtexteditor": "^20.1.57",
     "@techiediaries/ngx-qrcode": "^9.1.0",
     "angular2-qrcode": "^2.0.3",
@@ -39,7 +39,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~13.3.3",
     "@angular/cli": "~13.3.3",
-    "@angular/compiler-cli": "~13.3.0",
+    "@angular/compiler-cli": "~14.3.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~4.0.0",
