@@ -49,70 +49,63 @@ import {MatFormFieldModule} from '@angular/material/form-field';
 
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    PacientesComponent,
-    ConceptosComponent,
-    ConceptosFormComponent,
-    PacientesFormComponent,
-    VentaConceptosComponent,
-    BuscarEstudioModalComponent,
-    EnviarEstudioModalComponent,
-    SubirFotoOrdenComponent,
-    InformacionEstudioModalComponent,
-    AgendaComponent,
-    VerAgendadosModalComponent,
-    AntecedentesEstudioModalComponent,
-    MedicoRadiologoComponent,
-    DictadorComponent,
-    ResultadosComponent,
-    DictamenComponent,
-    EnviarEstudiosComponent,
-    EnviarEstudioComponent,
-    WorklistComponent,
-    SubirInterpretacionComponent,
-    OrdenVentaComponent,
-    LoginComponent
-  ],
-  entryComponents:[
-    BuscarEstudioModalComponent,
-    EnviarEstudioModalComponent,
-    InformacionEstudioModalComponent,
-    VerAgendadosModalComponent,
-    AntecedentesEstudioModalComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    LayoutModule,
-    HttpClientModule,
-    FormsModule,
-    BrowserAnimationsModule,
-    MatAutocompleteModule,
-    MatInputModule,
-    ReactiveFormsModule,
-    MatPaginatorModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatTableModule,
-    MatButtonModule,
-    MatDialogModule,
-    PdfViewerModule,
-    MatCheckboxModule,
-    MatSelectModule,
-    MatIconModule,
-    MatCardModule,
-    QRCodeModule,
-    RecepcionModule,
-    MatCardModule,
-    MatDividerModule,
-    MatIconModule,
-    MatSelectModule,
-    MatFormFieldModule
-  ],
-  providers: [
-    DatePipe,
-  ],
-  bootstrap: [AppComponent]
+    declarations: [
+        AppComponent,
+        PacientesComponent,
+        ConceptosComponent,
+        ConceptosFormComponent,
+        PacientesFormComponent,
+        VentaConceptosComponent,
+        BuscarEstudioModalComponent,
+        EnviarEstudioModalComponent,
+        SubirFotoOrdenComponent,
+        InformacionEstudioModalComponent,
+        AgendaComponent,
+        VerAgendadosModalComponent,
+        AntecedentesEstudioModalComponent,
+        MedicoRadiologoComponent,
+        DictadorComponent,
+        ResultadosComponent,
+        DictamenComponent,
+        EnviarEstudiosComponent,
+        EnviarEstudioComponent,
+        WorklistComponent,
+        SubirInterpretacionComponent,
+        OrdenVentaComponent,
+        LoginComponent
+    ],
+    imports: [
+        BrowserModule,
+        AppRoutingModule,
+        LayoutModule,
+        HttpClientModule,
+        FormsModule,
+        BrowserAnimationsModule,
+        MatAutocompleteModule,
+        MatInputModule,
+        ReactiveFormsModule,
+        MatPaginatorModule,
+        MatDatepickerModule,
+        MatNativeDateModule,
+        MatTableModule,
+        MatButtonModule,
+        MatDialogModule,
+        PdfViewerModule,
+        MatCheckboxModule,
+        MatSelectModule,
+        MatIconModule,
+        MatCardModule,
+        QRCodeModule,
+        RecepcionModule,
+        MatCardModule,
+        MatDividerModule,
+        MatIconModule,
+        MatSelectModule,
+        MatFormFieldModule
+    ],
+    providers: [
+        DatePipe,
+    ],
+    bootstrap: [AppComponent]
 })
 export class AppModule { }
